pyproject.toml
[project]
name = "code-buddy"
version = "0.1.0"
description = "REALM: A reverse engineering assistant that generates documentation from code"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "litellm>=1.70.0",
    "streamlit>=1.45.1",
    "chromadb>=0.4.22",
    "fastapi>=0.104.1",
    "uvicorn>=0.23.2",
    "python-dotenv>=1.0.0",
    "langchain>=0.0.335",
    "langchain-openai>=0.0.2",
    "tiktoken>=0.5.1",
    "pydantic>=2.4.2",
    "requests>=2.31.0",
    "sentence-transformers>=4.1.0",
]

[tool.setuptools]
packages = ["frontend", "backend", "shared"]
